requirements.txt
django-model-utils>=2.0
# Additional requirements go here
django-nested-admin==3.1.3
django-extensions==2.1.5
django-simple-history==2.7.0
